requirements.txt
sqlalchemy>=2.0
pyyaml>=6.0
requests>=2.31
